# openapi_test.json
{"openapi": "3.1.0", "info": {"title": "Shy2Ask API", "version": "1.0.0", "description": "**Auth:** register, login (email + OTP verification), forgot/reset password. **Profile:** GET/PATCH /profile/me, list users (staff). **Censor:** POST /censor/text and POST /censor/image (rule-based + OpenAI). **Requests:** Create and list shy requests, manage conversations. Protected routes use **Bearer token** (Header: `Authorization: Bearer <token>`)."}, "paths": {"/auth/register": {"post": {"operationId": "account_api_register", "summary": "Register", "parameters": [], "responses": {"201": {"description": "Created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterOut"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Register a new user. Email verification OTP is sent; verify with /auth/verify-email.", "tags": ["Auth"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterIn"}}}, "required": true}}}, "/auth/login": {"post": {"operationId": "account_api_login", "summary": "Login", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginOut"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}, "403": {"description": "Forbidden", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Login with email and password. Email must be verified first.", "tags": ["Auth"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginIn"}}}, "required": true}}}, "/auth/forgot-password": {"post": {"operationId": "account_api_forgot_password", "summary": "Forgot Password", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ForgotPasswordOut"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Send OTP to email for password reset. Always returns 200 to avoid email enumeration.", "tags": ["Auth"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ForgotPasswordIn"}}}, "required": true}}}, "/auth/reset-password": {"post": {"operationId": "account_api_reset_password", "summary": "Reset Password", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResetPasswordOut"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Reset password using email + OTP from forgot-password email.", "tags": ["Auth"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResetPasswordIn"}}}, "required": true}}}, "/auth/verify-email": {"post": {"operationId": "account_api_verify_email", "summary": "Verify Email", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VerifyEmailOut"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Verify email with OTP sent after register. Requires email + otp_code.", "tags": ["Auth"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/VerifyEmailIn"}}}, "required": true}}}, "/auth/resend-verification": {"post": {"operationId": "account_api_resend_verification", "summary": "Resend Verification", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResendVerificationOut"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Resend email verification OTP. Use if user did not receive or OTP expired.", "tags": ["Auth"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResendVerificationIn"}}}, "required": true}}}, "/profile/me": {"get": {"operationId": "account_api_profile_me", "summary": "Profile Me", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileOut"}}}}}, "description": "Get current user profile.", "tags": ["Profile"], "security": [{"AuthBearer": []}]}, "patch": {"operationId": "account_api_profile_update", "summary": "Profile Update", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileOut"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Update current user profile. Optionally upload profile_picture (multipart).", "tags": ["Profile"], "requestBody": {"content": {"multipart/form-data": {"schema": {"properties": {"profile_picture": {"anyOf": [{"format": "binary", "type": "string"}, {"type": "null"}], "title": "Profile Picture"}, "payload": {"$ref": "#/components/schemas/ProfileUpdateIn"}}, "title": "MultiPartBodyParams", "type": "object", "required": ["payload"]}}}, "required": true}, "security": [{"AuthBearer": []}]}}, "/profile/users": {"get": {"operationId": "account_api_list_users", "summary": "List Users", "parameters": [{"in": "query", "name": "limit", "schema": {"default": 20, "title": "Limit", "type": "integer"}, "required": false}, {"in": "query", "name": "offset", "schema": {"default": 0, "title": "Offset", "type": "integer"}, "required": false}, {"in": "query", "name": "search", "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Search"}, "required": false}], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedUsers"}}}}, "403": {"description": "Forbidden", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "List users with pagination and optional search (staff only).\nQuery: limit, offset, search (email, first_name, last_name, alias_name).", "tags": ["Profile"], "security": [{"AuthBearer": []}]}}, "/censor/text": {"post": {"operationId": "chat_censor_api_censor_text_endpoint", "summary": "Censor Text Endpoint", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TextCensorOut"}}}}}, "description": "Censor plain text. Uses DB terms (any language) + built-in lists + optional AI.\nReturns censored text, blocked flag, detected terms, and categories.", "tags": ["Censor"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TextCensorIn"}}}, "required": true}}}, "/censor/image": {"post": {"operationId": "chat_censor_api_censor_image_endpoint", "summary": "Censor Image Endpoint", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImageCensorOut"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Upload an image; engine runs OCR to extract text, then censors it.\nReturns censored extracted text, blocked flag, and detected terms.\nRequires Tesseract OCR installed (and pytesseract, Pillow) for OCR; otherwise returns ocr_available=False.", "tags": ["Censor"], "requestBody": {"content": {"multipart/form-data": {"schema": {"properties": {"image": {"format": "binary", "title": "Image", "type": "string"}}, "required": ["image"], "title": "FileParams", "type": "object"}}}, "required": true}}}, "/realtime/": {"post": {"operationId": "chat_realtime_api_create_request", "summary": "Create Request", "parameters": [], "responses": {"201": {"description": "Created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShyRequestOut"}}}}}, "description": "Create a new ShyRequest. Automatically creates a Conversation and sends notifications.", "tags": ["Requests"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShyRequestIn"}}}, "required": true}}, "get": {"operationId": "chat_realtime_api_list_requests", "summary": "List Requests", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ShyRequestOut"}, "title": "Response", "type": "array"}}}}}, "description": "List requests for the logged-in user.", "tags": ["Requests"], "security": [{"AuthBearer": []}]}}, "/realtime/{request_id}/conversation": {"get": {"operationId": "chat_realtime_api_get_conversation", "summary": "Get Conversation", "parameters": [{"in": "path", "name": "request_id", "schema": {"title": "Request Id", "type": "integer"}, "required": true}], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/MessageOut"}, "title": "Response", "type": "array"}}}}}, "description": "Retrieve messages for a given request.", "tags": ["Requests"]}}, "/realtime/{request_id}/messages": {"post": {"operationId": "chat_realtime_api_send_message", "summary": "Send Message", "parameters": [{"in": "path", "name": "request_id", "schema": {"title": "Request Id", "type": "integer"}, "required": true}], "responses": {"201": {"description": "Created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageOut"}}}}}, "description": "Send a message for a request (REST alternative to WebSocket).", "tags": ["Requests"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageIn"}}}, "required": true}}}, "/subscriptions": {"get": {"operationId": "chat_subscription_api_list_my_subscriptions", "summary": "List My Subscriptions", "parameters": [], "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/SubscriptionOut"}, "title": "Response", "type": "array"}}}}}, "description": "List current user's subscriptions (request updates, deal alerts).", "tags": ["Subscription"], "security": [{"AuthBearer": []}]}, "post": {"operationId": "chat_subscription_api_subscribe", "summary": "Subscribe", "parameters": [], "responses": {"201": {"description": "Created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SubscriptionOut"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Subscribe to request updates (for a request) or deal alerts (AI-detected deals).", "tags": ["Subscription"], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SubscribeIn"}}}, "required": true}, "security": [{"AuthBearer": []}]}}, "/subscriptions/{subscription_id}": {"delete": {"operationId": "chat_subscription_api_unsubscribe", "summary": "Unsubscribe", "parameters": [{"in": "path", "name": "subscription_id", "schema": {"title": "Subscription Id", "type": "integer"}, "required": true}], "responses": {"204": {"description": "No Content"}, "404": {"description": "Not Found", "content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response", "type": "object"}}}}}, "description": "Unsubscribe (deactivate) a subscription.", "tags": ["Subscription"], "security": [{"AuthBearer": []}]}}}, "components": {"schemas": {"RegisterOut": {"properties": {"id": {"title": "Id", "type": "integer"}, "email": {"title": "Email", "type": "string"}, "first_name": {"title": "First Name", "type": "string"}, "last_name": {"title": "Last Name", "type": "string"}, "alias_name": {"title": "Alias Name", "type": "string"}, "is_verified": {"title": "Is Verified", "type": "boolean"}, "token": {"title": "Token", "type": "string"}, "message": {"default": "Please verify your email with the OTP sent to your inbox.", "title": "Message", "type": "string"}}, "required": ["id", "email", "first_name", "last_name", "alias_name", "is_verified", "token"], "title": "RegisterOut", "type": "object"}, "RegisterIn": {"properties": {"email": {"title": "Email", "type": "string"}, "password": {"title": "Password", "type": "string"}, "first_name": {"default": "", "title": "First Name", "type": "string"}, "last_name": {"default": "", "title": "Last Name", "type": "string"}, "alias_name": {"default": "", "title": "Alias Name", "type": "string"}, "phone_number": {"default": "", "title": "Phone Number", "type": "string"}}, "required": ["email", "password"], "title": "RegisterIn", "type": "object"}, "LoginOut": {"properties": {"token": {"title": "Token", "type": "string"}, "user": {"additionalProperties": true, "title": "User", "type": "object"}}, "required": ["token", "user"], "title": "LoginOut", "type": "object"}, "LoginIn": {"properties": {"email": {"title": "Email", "type": "string"}, "password": {"title": "Password", "type": "string"}}, "required": ["email", "password"], "title": "LoginIn", "type": "object"}, "ForgotPasswordOut": {"properties": {"message": {"title": "Message", "type": "string"}}, "required": ["message"], "title": "ForgotPasswordOut", "type": "object"}, "ForgotPasswordIn": {"properties": {"email": {"title": "Email", "type": "string"}}, "required": ["email"], "title": "ForgotPasswordIn", "type": "object"}, "ResetPasswordOut": {"properties": {"message": {"title": "Message", "type": "string"}}, "required": ["message"], "title": "ResetPasswordOut", "type": "object"}, "ResetPasswordIn": {"properties": {"email": {"title": "Email", "type": "string"}, "otp": {"title": "Otp", "type": "string"}, "new_password": {"title": "New Password", "type": "string"}}, "required": ["email", "otp", "new_password"], "title": "ResetPasswordIn", "type": "object"}, "VerifyEmailOut": {"properties": {"message": {"title": "Message", "type": "string"}, "user": {"additionalProperties": true, "title": "User", "type": "object"}}, "required": ["message", "user"], "title": "VerifyEmailOut", "type": "object"}, "VerifyEmailIn": {"properties": {"email": {"title": "Email", "type": "string"}, "otp_code": {"title": "Otp Code", "type": "string"}}, "required": ["email", "otp_code"], "title": "VerifyEmailIn", "type": "object"}, "ResendVerificationOut": {"properties": {"message": {"title": "Message", "type": "string"}}, "required": ["message"], "title": "ResendVerificationOut", "type": "object"}, "ResendVerificationIn": {"properties": {"email": {"title": "Email", "type": "string"}}, "required": ["email"], "title": "ResendVerificationIn", "type": "object"}, "ProfileOut": {"properties": {"id": {"title": "Id", "type": "integer"}, "email": {"title": "Email", "type": "string"}, "first_name": {"title": "First Name", "type": "string"}, "last_name": {"title": "Last Name", "type": "string"}, "alias_name": {"title": "Alias Name", "type": "string"}, "phone_number": {"title": "Phone Number", "type": "string"}, "profile_picture": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Profile Picture"}, "is_verified": {"title": "Is Verified", "type": "boolean"}, "date_joined": {"title": "Date Joined", "type": "string"}, "updated_at": {"title": "Updated At", "type": "string"}}, "required": ["id", "email", "first_name", "last_name", "alias_name", "phone_number", "is_verified", "date_joined", "updated_at"], "title": "ProfileOut", "type": "object"}, "ProfileUpdateIn": {"properties": {"first_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "First Name"}, "last_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Name"}, "alias_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Alias Name"}, "phone_number": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Phone Number"}}, "title": "ProfileUpdateIn", "type": "object"}, "PaginatedUsers": {"properties": {"count": {"title": "Count", "type": "integer"}, "limit": {"title": "Limit", "type": "integer"}, "offset": {"title": "Offset", "type": "integer"}, "items": {"items": {"$ref": "#/components/schemas/UserListItem"}, "title": "Items", "type": "array"}}, "required": ["count", "limit", "offset", "items"], "title": "PaginatedUsers", "type": "object"}, "UserListItem": {"properties": {"id": {"title": "Id", "type": "integer"}, "email": {"title": "Email", "type": "string"}, "first_name": {"title": "First Name", "type": "string"}, "last_name": {"title": "Last Name", "type": "string"}, "alias_name": {"title": "Alias Name", "type": "string"}, "is_active": {"title": "Is Active", "type": "boolean"}, "date_joined": {"title": "Date Joined", "type": "string"}}, "required": ["id", "email", "first_name", "last_name", "alias_name", "is_active", "date_joined"], "title": "UserListItem", "type": "object"}, "TextCensorOut": {"properties": {"censored_text": {"title": "Censored Text", "type": "string"}, "blocked": {"title": "Blocked", "type": "boolean"}, "detected": {"items": {"additionalProperties": true, "type": "object"}, "title": "Detected", "type": "array"}, "categories": {"items": {"type": "string"}, "title": "Categories", "type": "array"}, "ai_toxic_score": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Ai Toxic Score"}, "ai_provider": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Ai Provider"}}, "required": ["censored_text", "blocked", "detected", "categories"], "title": "TextCensorOut", "type": "object"}, "TextCensorIn": {"properties": {"text": {"title": "Text", "type": "string"}}, "required": ["text"], "title": "TextCensorIn", "type": "object"}, "ImageCensorOut": {"properties": {"censored_text": {"title": "Censored Text", "type": "string"}, "blocked": {"title": "Blocked", "type": "boolean"}, "detected": {"items": {"additionalProperties": true, "type": "object"}, "title": "Detected", "type": "array"}, "categories": {"items": {"type": "string"}, "title": "Categories", "type": "array"}, "extracted_text": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Extracted Text"}, "ocr_available": {"title": "Ocr Available", "type": "boolean"}, "ai_toxic_score": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Ai Toxic Score"}, "ai_provider": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Ai Provider"}}, "required": ["censored_text", "blocked", "detected", "categories", "ocr_available"], "title": "ImageCensorOut", "type": "object"}, "AttachmentOut": {"properties": {"id": {"title": "Id", "type": "integer"}, "file": {"title": "File", "type": "string"}, "uploaded_at": {"format": "date-time", "title": "Uploaded At", "type": "string"}}, "required": ["id", "file", "uploaded_at"], "title": "AttachmentOut", "type": "object"}, "ShyRequestOut": {"properties": {"id": {"title": "Id", "type": "integer"}, "tracking_code": {"title": "Tracking Code", "type": "string"}, "requester_name": {"title": "Requester Name", "type": "string"}, "requester_email": {"title": "Requester Email", "type": "string"}, "requester_phone": {"title": "Requester Phone", "type": "string"}, "requester_alias": {"title": "Requester Alias", "type": "string"}, "target_name": {"title": "Target Name", "type": "string"}, "target_email": {"title": "Target Email", "type": "string"}, "target_phone": {"title": "Target Phone", "type": "string"}, "target_address": {"title": "Target Address", "type": "string"}, "description": {"title": "Description", "type": "string"}, "service_channel": {"title": "Service Channel", "type": "string"}, "call_minutes": {"title": "Call Minutes", "type": "integer"}, "quoted_price_chf": {"title": "Quoted Price Chf", "type": "number"}, "country_code": {"title": "Country Code", "type": "string"}, "status": {"title": "Status", "type": "string"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "attachments": {"default": [], "items": {"$ref": "#/components/schemas/AttachmentOut"}, "title": "Attachments", "type": "array"}}, "required": ["id", "tracking_code", "requester_name", "requester_email", "requester_phone", "requester_alias", "target_name", "target_email", "target_phone", "target_address", "description", "service_channel", "call_minutes", "quoted_price_chf", "country_code", "status", "created_at"], "title": "ShyRequestOut", "type": "object"}, "ShyRequestIn": {"properties": {"requester_name": {"title": "Requester Name", "type": "string"}, "requester_email": {"title": "Requester Email", "type": "string"}, "requester_phone": {"default": "", "title": "Requester Phone", "type": "string"}, "requester_alias": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Requester Alias"}, "target_name": {"default": "", "title": "Target Name", "type": "string"}, "target_email": {"default": "", "title": "Target Email", "type": "string"}, "target_phone": {"default": "", "title": "Target Phone", "type": "string"}, "target_address": {"default": "", "title": "Target Address", "type": "string"}, "description": {"title": "Description", "type": "string"}, "service_channel": {"default": "email", "title": "Service Channel", "type": "string"}, "call_minutes": {"default": 0, "title": "Call Minutes", "type": "integer"}}, "required": ["requester_name", "requester_email", "description"], "title": "ShyRequestIn", "type": "object"}, "MessageOut": {"properties": {"id": {"title": "Id", "type": "integer"}, "sender": {"title": "Sender", "type": "string"}, "sender_display_name": {"title": "Sender Display Name", "type": "string"}, "display_name": {"title": "Display Name", "type": "string"}, "body": {"title": "Body", "type": "string"}, "clean_body": {"title": "Clean Body", "type": "string"}, "is_blocked": {"title": "Is Blocked", "type": "boolean"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}}, "required": ["id", "sender", "sender_display_name", "display_name", "body", "clean_body", "is_blocked", "created_at"], "title": "MessageOut", "type": "object"}, "MessageIn": {"properties": {"body": {"title": "Body", "type": "string"}, "alias": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Alias"}}, "required": ["body"], "title": "MessageIn", "type": "object"}, "SubscriptionOut": {"properties": {"id": {"title": "Id", "type": "integer"}, "subscription_type": {"title": "Subscription Type", "type": "string"}, "request_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Request Id"}, "tracking_code": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Tracking Code"}, "is_active": {"title": "Is Active", "type": "boolean"}, "created_at": {"title": "Created At", "type": "string"}}, "required": ["id", "subscription_type", "request_id", "tracking_code", "is_active", "created_at"], "title": "SubscriptionOut", "type": "object"}, "SubscribeIn": {"properties": {"subscription_type": {"title": "Subscription Type", "type": "string"}, "request_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Request Id"}}, "required": ["subscription_type"], "title": "SubscribeIn", "type": "object"}}, "securitySchemes": {"AuthBearer": {"type": "http", "scheme": "bearer"}}}, "servers": [], "tags": [{"name": "Auth", "description": "Register, login, verify email, password reset"}, {"name": "Profile", "description": "Current user profile and staff user list"}, {"name": "Censor", "description": "Text and image content moderation"}, {"name": "Requests", "description": "Real-time ShyRequests and Conversations"}]}